% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgumbel.R
\name{dbgumbel}
\alias{dbgumbel}
\title{Bimodal Gumbel: Density Function}
\usage{
dbgumbel(x, mu, sigma, delta)
}
\arguments{
\item{x}{Domain.}

\item{mu}{First location parameter.}

\item{sigma}{Scale parameter.}

\item{delta}{Second location parameter.}
}
\description{
Bimodal Gumbel: Density Function
}
\examples{
# dbgumbel(x = 0, mu = -2, sigma = 1, delta = -1)
# curve(dbgumbel(x, mu = -2, sigma = 1, delta = -1), xlim = c(-5, 10))
# integrate(dbgumbel, mu = -2, sigma = 1, delta = -1, lower = -5, upper = 0)
}
