% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1bgumbel.R
\name{m1bgumbel}
\alias{m1bgumbel}
\title{Bimodal Gumbel: Theoretical E(X)}
\usage{
m1bgumbel(mu, sigma, delta)
}
\arguments{
\item{mu}{First location parameter.}

\item{sigma}{Scale parameter.}

\item{delta}{Second location parameter.}
}
\description{
Bimodal Gumbel: Theoretical E(X)
}
